requirements.txt
# FAB - Firewall Access Bot dependencies

# Telegram Bot
python-telegram-bot==21.0.1

# Web Framework
Flask==3.0.3
Werkzeug==3.0.4

# Message Queue
pika==1.3.2

# Environment Variables
python-dotenv==1.0.1

# Data Validation
pydantic==2.7.4

# Development dependencies
black==24.4.2
flake8==7.1.0
mypy==1.10.1
pytest==8.2.2

